refactor(asset): modularise asset/library
